fix(statistics): No renderiza correctamente la pantalla
